IGNITE-7876: Remove unused code
