+++ modules/ml/src/test/java/org/apache/ignite/ml/svm/SVMTestSuite.java
@@ -17,17 +17,16 @@
 
 package org.apache.ignite.ml.svm;
 
-import org.apache.ignite.ml.svm.multi.DistributedLinearSVMMultiClassClassificationTrainerTest;
-import org.apache.ignite.ml.svm.multi.LocalLinearSVMMultiClassClassificationTrainerTest;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 
 /**
- * Test suite for all tests located in org.apache.ignite.ml.regressions.* package.
+ * Test suite for all tests located in org.apache.ignite.ml.svm.* package.
  */
 @RunWith(Suite.class)
 @Suite.SuiteClasses({
-    SVMModelTest.class
+    SVMModelTest.class,
+    SVMBinaryTrainerTest.class
 })
 public class SVMTestSuite {
     // No-op.
